requirements.txt
fastapi
python-dotenv
langchain
openai
redis
convokit
pandas
numpy
pydantic
pyairtable
airtable
huggingface-hub
uvicorn
